商品查询

+++ youpin-system/src/main/java/com/hflw/vasp/modules/entity/GoodsPicture.java
@@ -23,7 +23,7 @@ public class GoodsPicture extends BaseEntity {
     /**
      * 商品id
      */
-    private Integer goodsId;
+    private Long goodsId;
     /**
      * 图片url
      */
